backend/requirements.txt
fastapi
uvicorn
psycopg2-binarypython-dotenv
